requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
PyYAML>=6.0

# Data and backtesting
yfinance>=0.2.0
vectorbt>=0.25.0
pandas-ta>=0.3.14b

# Reporting (pyfolio optional for PDF generation)
pyfolio>=0.9.2
matplotlib>=3.5.0
seaborn>=0.11.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
mypy>=1.0.0

# Optimization dependencies
tqdm>=4.64.0  # Progress indicators for optimization
scipy>=1.9.0  # Statistical analysis and constraint evaluation
